Move WaterYearDto to its own file, add to parcel update layer initial data grab
