+++ Source/Rio.Models/DataTransferObjects/WaterYearDto.cs
@@ -7,5 +7,6 @@ namespace Rio.Models.DataTransferObjects
         public int WaterYearID { get; set; }
         public int Year { get; set; }
         public DateTime? FinalizeDate { get; set; }
+        public DateTime? ParcelLayerUpdateDate { get; set; }
     }
 }
