+++ Source/Rio.Web/src/app/pages/landowner-dashboard/landowner-dashboard.component.ts
@@ -24,7 +24,7 @@ import { environment } from 'src/environments/environment';
 import { LandownerWaterUseChartComponent } from '../landowner-water-use-chart/landowner-water-use-chart.component';
 import { ParcelAllocationTypeDto } from 'src/app/shared/models/parcel-allocation-type-dto';
 import { ParcelAllocationTypeService } from 'src/app/services/parcel-allocation-type.service';
-import { WaterYearDto } from 'src/app/shared/models/openet-sync-history-dto';
+import { WaterYearDto } from "src/app/shared/models/water-year-dto";
 import { WaterYearService } from 'src/app/services/water-year.service';
 
 @Component({
