+++ Source/Rio.Web/src/app/pages/manager-dashboard/manager-dashboard.component.ts
@@ -25,7 +25,7 @@ import { MultiSeriesEntry, SeriesEntry } from 'src/app/shared/models/series-entr
 import { AccountService } from 'src/app/services/account/account.service';
 import { WaterAllocationOverviewDto } from 'src/app/shared/models/water-usage-dto';
 import { WaterYearService } from 'src/app/services/water-year.service';
-import { WaterYearDto } from 'src/app/shared/models/openet-sync-history-dto';
+import { WaterYearDto } from "src/app/shared/models/water-year-dto";
 
 @Component({
   selector: 'rio-manager-dashboard',
