+++ Source/Rio.Web/src/app/pages/openet-sync-water-year-status-list/openet-sync-water-year-status-list.component.ts
@@ -10,7 +10,8 @@ import { UserDto } from 'src/app/shared/models';
 import { Alert } from 'src/app/shared/models/alert';
 import { AlertContext } from 'src/app/shared/models/enums/alert-context.enum';
 import { CustomRichTextType } from 'src/app/shared/models/enums/custom-rich-text-type.enum';
-import { OpenETSyncHistoryDto, WaterYearDto } from 'src/app/shared/models/openet-sync-history-dto';
+import { OpenETSyncHistoryDto } from 'src/app/shared/models/openet-sync-history-dto';
+import { WaterYearDto } from "src/app/shared/models/water-year-dto";
 import { OpenETSyncResultTypeEnum } from 'src/app/shared/models/openet-sync-result-type-dto';
 import { WaterYearQuickOpenETHistoryDto } from 'src/app/shared/models/water-year-quick-open-et-history-dto';
 import { AlertService } from 'src/app/shared/services/alert.service';
