+++ Source/Rio.Web/src/app/pages/parcel-override-et-data/parcel-override-et-data.component.ts
@@ -11,7 +11,7 @@ import { AccountDto } from 'src/app/shared/models/account/account-dto';
 import { AlertService } from 'src/app/shared/services/alert.service';
 import { Alert } from 'src/app/shared/models/alert';
 import { AlertContext } from 'src/app/shared/models/enums/alert-context.enum';
-import { WaterYearDto } from 'src/app/shared/models/openet-sync-history-dto';
+import { WaterYearDto } from "src/app/shared/models/water-year-dto";
 import { WaterYearService } from 'src/app/services/water-year.service';
