+++ Source/Rio.Web/src/app/shared/models/openet-sync-history-dto.ts
@@ -1,4 +1,5 @@
 import { OpenETSyncResultTypeDto } from "./openet-sync-result-type-dto";
+import { WaterYearDto } from './water-year-dto';
 
 
 export class OpenETSyncHistoryDto {
@@ -13,12 +14,4 @@ export class OpenETSyncHistoryDto {
     }
 }
 
-export class WaterYearDto {
-    WaterYearID: number;
-    Year: number;
-    FinalizeDate: Date;
 
-    constructor(obj?: any) {
-        Object.assign(this, obj);
-    }
-}
