+++ Source/Rio.Web/src/app/shared/models/water-year-dto.ts
@@ -1,0 +1,10 @@
+
+export class WaterYearDto {
+    WaterYearID: number;
+    Year: number;
+    FinalizeDate: Date;
+
+    constructor(obj?: any) {
+        Object.assign(this, obj);
+    }
+}
