+++ Source/Rio.Web/src/app/shared/models/water-year-quick-open-et-history-dto.ts
@@ -1,4 +1,4 @@
-import { WaterYearDto } from './openet-sync-history-dto';
+import { WaterYearDto } from "./water-year-dto";
 
 export class WaterYearQuickOpenETHistoryDto {
     WaterYear: WaterYearDto;
